+++ budiga_app/MVVM/Model/ItemModel.cs
@@ -15,7 +15,7 @@ namespace budiga_app.MVVM.Model
         private string _barcode;
         private string _name;
         private string _brand;
-        private int _price;
+        private float _price;
         private int _quantity;
         private bool _isDeleted;
         private ObservableCollection<ItemModel> _itemRecords;
@@ -25,7 +25,7 @@ namespace budiga_app.MVVM.Model
         public string Barcode { get { return _barcode; } set { _barcode = value; OnPropertyChanged("Barcode"); } }
         public string Name { get { return _name; } set { _name = value; OnPropertyChanged("Name");} }
         public string Brand { get { return _brand; } set { _brand = value; OnPropertyChanged("Brand"); } }
-        public int Price { get { return _price; } set { _price = value; OnPropertyChanged("Price"); } }
+        public float Price { get { return _price; } set { _price = value; OnPropertyChanged("Price"); } }
         public int Quantity { get { return _quantity; } set { _quantity = value; OnPropertyChanged("Quantity"); } }
         public bool IsDeleted { get { return _isDeleted; } set { _isDeleted = value; OnPropertyChanged("IsDeleted"); } }
         public ObservableCollection<ItemModel> ItemRecords { get { return _itemRecords; } set { _itemRecords = value; OnPropertyChanged("ItemRecords"); } }
